requirements.txt
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.11.5
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.4.1
pycodestyle==2.5.0
pycparser==2.19
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
